requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
openapi-codec==1.3.2
Pillow==8.1.2
psycopg2 @ file:///C:/ci/psycopg2_1612298715048/work
pyglet==1.5.14
PyJWT==1.7.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.25.1
simplejson==3.17.2
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
uritemplate==3.0.1
urllib3==1.26.4
wincertstore==0.2
